Fix test build
