+++ src/Package/Impl/Shell/VsAppShell.Services.cs
@@ -3,9 +3,7 @@
 
 using System.IO;
 using Microsoft.Common.Core.Imaging;
-using Microsoft.Common.Core.IO;
 using Microsoft.Common.Core.Logging;
-using Microsoft.Common.Core.OS;
 using Microsoft.Common.Core.Security;
 using Microsoft.Common.Core.Services;
 using Microsoft.Common.Core.Shell;
@@ -20,15 +18,11 @@ using Microsoft.R.Components;
 using Microsoft.R.Components.Security;
 using Microsoft.R.Components.Settings;
 using Microsoft.R.Components.StatusBar;
-using Microsoft.R.Containers;
 using Microsoft.R.Editor;
 using Microsoft.R.Editor.Settings;
 using Microsoft.R.Host.Client;
 using Microsoft.R.Platform;
 using Microsoft.R.Platform.Interpreters;
-using Microsoft.R.Platform.IO;
-using Microsoft.R.Platform.Logging;
-using Microsoft.R.Platform.OS;
 using Microsoft.VisualStudio.ComponentModelHost;
 using Microsoft.VisualStudio.R.Package.Imaging;
 using Microsoft.VisualStudio.R.Package.Options.R;
