--- src/UnitTests/Windows/Impl/Microsoft.UnitTests.Core.Windows.csproj
@@ -1,97 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{80B1E727-AD90-4CA8-934B-B9D3264547E8}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Microsoft.UnitTests.Core</RootNamespace>
-    <AssemblyName>Microsoft.UnitTests.Core.Windows</AssemblyName>
-    <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
-    <FileAlignment>512</FileAlignment>
-    <SccProjectName>SAK</SccProjectName>
-    <SccLocalPath>SAK</SccLocalPath>
-    <SccAuxPath>SAK</SccAuxPath>
-    <SccProvider>SAK</SccProvider>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'">
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|AnyCPU'">
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <Import Project="..\..\..\R.Settings.props" />
-  <ItemGroup>
-    <Compile Include="$(SourceDirectory)AssemblyVersionInfo.cs">
-      <Link>Properties\AssemblyVersionInfo.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\GlobalAssemblyInfo.cs">
-      <Link>Properties\GlobalAssemblyInfo.cs</Link>
-    </Compile>
-    <Compile Include="XUnit\AssemblyLoaderAttribute.cs" />
-    <Compile Include="XUnit\ClassRunner.cs" />
-    <Compile Include="XUnit\TaskObserver.cs" />
-    <Compile Include="XUnit\TestCase.cs" />
-    <Compile Include="XUnit\TestMethodRunner.cs" />
-    <Compile Include="Threading\TestMainThread.cs" />
-    <Compile Include="XUnit\MessageBusInjections\VerifyGlobalProviderMessageBusInjection.cs" />
-    <Compile Include="XUnit\MethodFixtures\ContainerHostMethodFixture.cs" />
-    <Compile Include="XUnit\TestTraceListener.cs" />
-    <Compile Include="XUnit\VsAssemblyLoaderAttribute.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="project.json" />
-  </ItemGroup>
-  <ItemGroup>
-    <Reference Include="PresentationCore" />
-    <Reference Include="PresentationFramework" />
-    <Reference Include="System" />
-    <Reference Include="System.ComponentModel.Composition" />
-    <Reference Include="System.Drawing" />
-    <Reference Include="System.Windows.Forms" />
-    <Reference Include="System.Xaml" />
-    <Reference Include="WindowsBase" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="EventTaskSources.cs" />
-    <Compile Include="Mef\CatalogFactory.cs" />
-    <Compile Include="Mef\PartMetadataAttributeNames.cs" />
-    <Compile Include="NativeMethods.cs" />
-    <Compile Include="NSubstitute\Mef\MefSubstituteBuilder.AddValueFactory.cs" />
-    <Compile Include="NSubstitute\Mef\ImportDefinitionConstraintAnalyser.cs" />
-    <Compile Include="NSubstitute\Mef\MefSubstitute.cs" />
-    <Compile Include="NSubstitute\Mef\MefSubstituteBuilder.cs" />
-    <Compile Include="NSubstitute\Mef\NSubstituteCompositionContainer.cs" />
-    <Compile Include="NSubstitute\Mef\NSubstituteExportProvider.cs" />
-    <Compile Include="Paths.cs" />
-    <Compile Include="ProcessUtils.cs" />
-    <Compile Include="Random.cs" />
-    <Compile Include="Threading\UIThreadHelper.cs" />
-    <Compile Include="Threading\UIThreadTools.cs" />
-    <Compile Include="UI\ContainerHost.cs" />
-    <Compile Include="Mef\IExportProvider.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\Common\Core\Impl\Microsoft.R.Common.Core.csproj">
-      <Project>{C2A86AAB-A537-418D-A1EF-14540D340CB1}</Project>
-      <Name>Microsoft.R.Common.Core</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\R\Platform\Impl\Microsoft.R.Platform.csproj">
-      <Project>{9000d374-e37e-4df2-a703-08d29f3fbe61}</Project>
-      <Name>Microsoft.R.Platform</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\Windows\R\Platform\Impl\Microsoft.R.Platform.Windows.csproj">
-      <Project>{de1ec50a-c588-45a9-b22c-981419cd7cd4}</Project>
-      <Name>Microsoft.R.Platform.Windows</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\Core\Impl\Microsoft.UnitTests.Core.csproj">
-      <Project>{5ef2ad64-d6fe-446b-b350-8c7f0df0834d}</Project>
-      <Name>Microsoft.UnitTests.Core</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <Import Project="$(SourceDirectory)R.Build.Version.targets" />
-  <Import Project="$(SourceDirectory)R.Build.Sign.targets" />
-</Project>
