--- src/UnitTests/Windows/Impl/project.json
@@ -1,18 +1,0 @@
-﻿{
-  "dependencies": {
-    "FluentAssertions": "4.19.3",
-    "MicroBuild.Core": "0.2.0",
-    "Microsoft.VisualStudio.Shell.15.0": "15.4.27004",
-    "NSubstitute": "2.0.3",
-    "System.Threading.Tasks.Dataflow": "4.8.0",
-    "System.ValueTuple": "4.4.0",
-    "xunit": "2.3.1",
-    "xunit.runner.visualstudio": "2.3.1"
-  },
-  "frameworks": {
-    "net461": {}
-  },
-  "runtimes": {
-    "win": {}
-  }
-}
