+++ src/Windows/Containers/Test/WindowsContainerServiceTest.cs
@@ -16,7 +16,9 @@ using Microsoft.R.Containers.Docker;
 using Microsoft.UnitTests.Core.XUnit;
 using System.IO;
 using System;
-using System.Collections.Generic;
+using Microsoft.R.Platform.IO;
+using Microsoft.R.Platform.OS;
+using NSubstitute;
 
 namespace Microsoft.R.Containers.Windows.Test {
     [ExcludeFromCodeCoverage]
