+++ src/Windows/R/Components/Test/Fixtures/RComponentServicesFixture.cs
@@ -31,7 +31,6 @@ namespace Microsoft.R.Components.Test.Fixtures {
             "Microsoft.VisualStudio.Platform.VSEditor.dll",
             "Microsoft.R.Components.dll",
             "Microsoft.R.Components.Windows.dll",
-            "Microsoft.R.Components.Test.dll",
             "Microsoft.R.Components.Windows.Test.dll"
         };
