+++ src/Windows/R/Editor/Application.Test/Completion/IntellisenseTest.cs
@@ -255,8 +255,7 @@ namespace Microsoft.R.Editor.Application.Test.Completion {
             }
         }
 
-        // Disabled since auto-insertion of braces is off
-        //[Test]
+        [Test(Skip = "Disabled since auto-insertion of braces is off")]
         [Category.Interactive]
         public async Task R_CompletionFunctionBraces01() {
             using (var script = await _editorHost.StartScript(Services, RContentTypeDefinition.ContentType, Workflow.RSessions)) {
