+++ src/Windows/R/Platform.Core/Test/Properties/AssemblyInfo.cs
